Abstracted logic for processing block traces
- We need this for handling slight differences between `mpt` & `smt`
  payloads.
