+++ trace_decoder/src/processed_block_trace_smt.rs
@@ -1,12 +1,17 @@
+use std::collections::HashMap;
+
 use ethereum_types::{Address, U256};
+use evm_arithmetization_mpt::generation::mpt::AccountRlp;
 
 use crate::{
     aliased_crate_types::SmtGenerationInputs,
-    decoding_smt::SmtTraceParsingResult,
-    processed_block_trace::ProcessedBlockTrace,
-    processed_block_trace_mpt::ProcessingMeta,
-    trace_protocol::BlockTrace,
-    types::{CodeHashResolveFunc, OtherBlockData},
+    decoding::TraceDecodingResult,
+    processed_block_trace::{
+        BlockTraceProcessing, ProcessedBlockTrace, ProcessedSectionInfo, ProcessingMeta,
+    },
+    protocol_processing::TraceProtocolDecodingResult,
+    trace_protocol::{BlockTrace, TriePreImage},
+    types::{CodeHash, CodeHashResolveFunc, HashedAccountAddr, OtherBlockData},
 };
 
 pub(crate) type SmtProcessedBlockTrace = ProcessedBlockTrace<ProcedBlockTraceSmtSpec>;
@@ -18,12 +23,46 @@ pub struct SmtProcessedBlockTracePreImage {}
 #[derive(Debug)]
 pub(crate) struct ProcedBlockTraceSmtSpec {}
 
+struct SmtBlockTraceProcessing;
+
+impl BlockTraceProcessing for SmtBlockTraceProcessing {
+    type ProcessedPreImage = SmtProcessedBlockTracePreImage;
+    type Output = ProcedBlockTraceSmtSpec;
+
+    fn process_block_trace(
+        image: TriePreImage,
+    ) -> TraceProtocolDecodingResult<Self::ProcessedPreImage> {
+        todo!()
+    }
+
+    fn get_account_keys(
+        image: &Self::ProcessedPreImage,
+    ) -> impl Iterator<Item = (HashedAccountAddr, AccountRlp)> {
+        todo!();
+
+        std::iter::empty()
+    }
+
+    fn get_any_extra_code_hash_mappings(
+        image: &Self::ProcessedPreImage,
+    ) -> Option<&HashMap<CodeHash, Vec<u8>>> {
+        todo!()
+    }
+
+    fn create_spec_output(
+        image: Self::ProcessedPreImage,
+        sect_info: ProcessedSectionInfo,
+    ) -> Self::Output {
+        todo!()
+    }
+}
+
 impl BlockTrace {
     /// Process the block trace into SMT IR.
     pub fn into_proof_gen_smt_ir(
         self,
         _other_data: OtherBlockData,
-    ) -> SmtTraceParsingResult<Vec<SmtGenerationInputs>> {
+    ) -> TraceDecodingResult<Vec<SmtGenerationInputs>> {
         todo!()
     }
 
@@ -33,26 +72,26 @@ impl BlockTrace {
         self,
         p_meta: &ProcessingMeta<F>,
         other_data: OtherBlockData,
-    ) -> SmtTraceParsingResult<Vec<SmtGenerationInputs>>
+    ) -> TraceProtocolDecodingResult<Vec<SmtGenerationInputs>>
     where
         F: CodeHashResolveFunc,
     {
         let processed_block_trace =
             self.into_smt_processed_block_trace(p_meta, other_data.b_data.withdrawals.clone())?;
 
-        processed_block_trace.into_proof_gen_ir(other_data)
+        let res = processed_block_trace.into_proof_gen_ir(other_data)?;
+
+        Ok(res)
     }
 
     fn into_smt_processed_block_trace<F>(
         self,
         p_meta: &ProcessingMeta<F>,
         withdrawals: Vec<(Address, U256)>,
-    ) -> SmtTraceParsingResult<SmtProcessedBlockTrace>
+    ) -> TraceProtocolDecodingResult<SmtProcessedBlockTrace>
     where
         F: CodeHashResolveFunc,
     {
-        let _ = p_meta;
-        let _ = withdrawals;
-        todo!()
+        self.into_processed_block_trace::<_, SmtBlockTraceProcessing>(p_meta, withdrawals)
     }
 }
